requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
astroid==2.3.3
Django==2.2.5
django-filter==2.2.0
django-graphql-playground==1.0
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
promise==2.3
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
utils==1.0.1
wrapt==1.11.2
